third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFEDisplacementMapElement-dom-in2-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'in2' attribute of the SVGFEDisplacementMapElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS displacementMap.getAttribute('in2') is "Texture"
PASS displacementMap.getAttribute('in2') is "Map"
PASS successfullyParsed is true

TEST COMPLETE
